+++ docs/source/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - :func:`_Table.exists` correctly ignores the partition decorator in case it is provided
 - :func:`_Table.create` now supports the possibility to pass extra table resource details, allowing partitioned table creationsphinx-quickstart
 - :func:`read_gbq` now raises ``QueryTimeout`` if the request exceeds the ``query.timeoutMs`` value specified in the BigQuery configuration. (:issue:`76`)
+- :func:`_Dataset.delete` now supports the delete_contents parameter
 
 0.2.0 / 2017-07-24
 ------------------